discussion fixes
